+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Utils Changelog
 
+## v0.0.6
+
+* Added `HasChanges` to `external_git`.
+* Added `SetPullRequestAutoComplete` to `azuredevops`.
+* Added `wiki` to `azuredevops`.
+* Added `identity` to `azuredevops`.
+* Added create and init repository to `azuredevops`.
+* Added `getProjectUUID` to `azuredevops`.
+
+## v0.0.5
+
+* Improved error handling in `azuredevops` package.
+
 ## v0.0.4
 
 * Updated `azuredevops` package to include pull request support.